requirements.txt
# AI IR Agent Dependencies

# Claude API
anthropic>=0.18.0

# RAG and Embeddings (local, no OpenAI needed)
sentence-transformers>=2.2.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10

# SSH Execution
paramiko>=3.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
click>=8.0.0

# Web scraping for corpus building
requests>=2.28.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# Testing
pytest>=7.0.0
